switch back to noop, start local packages
