+++ robot/packages/deferred_package_manager_test.go
@@ -99,7 +99,7 @@ func TestDeferredPackageManager(t *testing.T) {
 		// Assert that the cloud manager is nil initially
 		mgr, err := bag.pm.getManagerForSync(bag.ctx, []config.PackageConfig{})
 		test.That(t, err, test.ShouldBeNil)
-		_, isNoop := mgr.(*localManager)
+		_, isNoop := mgr.(*noopManager)
 		test.That(t, isNoop, test.ShouldBeTrue)
 		// send a msg on the chan indicating a connection
 		bag.mockChan <- mockChanVal{client: bag.client, err: nil}
@@ -177,7 +177,7 @@ func TestDeferredPackageManager(t *testing.T) {
 
 		err := bag.pm.Sync(bag.ctx, []config.PackageConfig{}, []config.Module{})
 		test.That(t, err, test.ShouldBeNil)
-		_, isNoop := bag.pm.lastSyncedManager.(*localManager)
+		_, isNoop := bag.pm.lastSyncedManager.(*noopManager)
 		test.That(t, isNoop, test.ShouldBeTrue)
 
 		// send a msg on the chan indicating a connection
